Add tests to, and move, docs class
