+++ docs/rst/manual/reference/forward_models.rst
@@ -1,3 +1,4 @@
 
 .. ert_forward_models::
+
     These are the forward models installed in your ERT environment.
